requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
blinker==1.9.0
build==1.2.2.post1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
Flask==3.1.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
Jinja2==3.1.5
jwcrypto==1.5.6
jwt==1.3.1
keyring==25.6.0
MarkupSafe==3.0.2
more-itertools==10.6.0
packaging==24.2
pluggy==1.5.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
pyproject_hooks==1.2.0
pytest==8.3.4
requests==2.32.3
SecretStorage==3.3.3
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
